Refactor code structure and expose `Config`
- Move `AgentManager` and `CertManager` traits from the graphql module
  to a new backend module.
- Add `Config` type to the public API under the backend module to ensure
  accessibility of all related types.

This reorganization improves code clarity by aligning the module
structure with the functionality usage.
