+++ src/graphql.rs
@@ -44,16 +44,14 @@ pub use self::sampling::{
     Policy as SamplingPolicy,
 };
 pub use self::trusted_user_agent::get_trusted_user_agent_list;
-use anyhow::anyhow;
+use crate::backend::{AgentManager, CertManager};
 use async_graphql::connection::ConnectionNameType;
 use async_graphql::{
     connection::{Connection, Edge, EmptyFields},
     Context, Guard, MergedObject, MergedSubscription, ObjectType, OutputType, Result,
 };
-use async_trait::async_trait;
 use chrono::TimeDelta;
 use data_encoding::BASE64;
-use ipnet::IpNet;
 use num_traits::ToPrimitive;
 use review_database::{
     self as database, types::FromKeyValue, Database, Direction, IterableMap, Role, Store,
@@ -61,8 +59,6 @@ use review_database::{
 pub use roxy::{Process, ResourceUsage};
 use std::{
     cmp,
-    collections::HashMap,
-    path::PathBuf,
     sync::{Arc, Mutex},
 };
 use tokio::sync::{Notify, RwLock};
@@ -72,109 +68,7 @@ use vinum::signal;
 /// GraphQL schema type.
 pub(super) type Schema = async_graphql::Schema<Query, Mutation, Subscription>;
 
-#[async_trait]
-pub trait AgentManager: Send + Sync {
-    async fn broadcast_trusted_domains(&self) -> Result<(), anyhow::Error> {
-        Err(anyhow!("Not supported"))
-    }
-
-    async fn broadcast_internal_networks(
-        &self,
-        _networks: &[u8],
-    ) -> Result<Vec<String>, anyhow::Error>;
-
-    async fn broadcast_allow_networks(
-        &self,
-        _networks: &[u8],
-    ) -> Result<Vec<String>, anyhow::Error>;
-
-    async fn broadcast_block_networks(
-        &self,
-        _networks: &[u8],
-    ) -> Result<Vec<String>, anyhow::Error>;
-
-    async fn broadcast_trusted_user_agent_list(&self, _list: &[u8]) -> Result<(), anyhow::Error> {
-        Err(anyhow!("Not supported"))
-    }
-
-    async fn online_apps_by_host_id(
-        &self,
-    ) -> Result<HashMap<String, Vec<(String, String)>>, anyhow::Error>; // (hostname, (agent_key, app_name))
-
-    async fn broadcast_crusher_sampling_policy(
-        &self,
-        _sampling_policies: &[SamplingPolicy],
-    ) -> Result<(), anyhow::Error>;
-
-    /// Returns the configuration of the given agent.
-    async fn get_config(
-        &self,
-        _hostname: &str,
-        _agent_id: &str,
-    ) -> Result<review_protocol::types::Config, anyhow::Error>;
-
-    /// Returns the list of processes running on the given host.
-    async fn get_process_list(&self, _hostname: &str) -> Result<Vec<Process>, anyhow::Error>;
-
-    /// Returns the resource usage of the given host.
-    async fn get_resource_usage(&self, _hostname: &str) -> Result<ResourceUsage, anyhow::Error>;
-
-    /// Halts the node with the given hostname.
-    async fn halt(&self, _hostname: &str) -> Result<(), anyhow::Error>;
-
-    /// Sends a ping message to the given host and waits for a response. Returns
-    /// the round-trip time in microseconds.
-    async fn ping(&self, _hostname: &str) -> Result<i64, anyhow::Error>;
-
-    /// Reboots the node with the given hostname.
-    async fn reboot(&self, _hostname: &str) -> Result<(), anyhow::Error>;
-
-    /// Sets the configuration of the given agent.
-    async fn set_config(
-        &self,
-        _hostname: &str,
-        _agent_id: &str,
-        _config: &review_protocol::types::Config,
-    ) -> Result<(), anyhow::Error>;
-
-    /// Updates the traffic filter rules for the given host.
-    async fn update_traffic_filter_rules(
-        &self,
-        _host: &str,
-        _rules: &[(IpNet, Option<Vec<u16>>, Option<Vec<u16>>)],
-    ) -> Result<(), anyhow::Error> {
-        Err(anyhow!("Not supported"))
-    }
-}
-
 type BoxedAgentManager = Box<dyn AgentManager>;
-
-pub trait CertManager: Send + Sync {
-    /// Returns the certificate path.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the certificate path cannot be determined.
-    fn cert_path(&self) -> Result<PathBuf, anyhow::Error>;
-
-    /// Returns the key path.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the key path cannot be determined.
-    fn key_path(&self) -> Result<PathBuf, anyhow::Error>;
-
-    /// Updates the certificate and key.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the certificate and key cannot be updated.
-    fn update_certificate(
-        &self,
-        cert: String,
-        key: String,
-    ) -> Result<Vec<ParsedCertificate>, anyhow::Error>;
-}
 
 /// Builds a GraphQL schema with the given database connection pool as its
 /// context.
@@ -720,7 +614,7 @@ fn validate_and_process_pagination_params(
 struct MockAgentManager {}
 
 #[cfg(test)]
-#[async_trait]
+#[async_trait::async_trait]
 impl AgentManager for MockAgentManager {
     async fn broadcast_internal_networks(
         &self,
@@ -746,8 +640,8 @@ impl AgentManager for MockAgentManager {
     }
     async fn online_apps_by_host_id(
         &self,
-    ) -> Result<HashMap<String, Vec<(String, String)>>, anyhow::Error> {
-        Ok(HashMap::new())
+    ) -> Result<std::collections::HashMap<String, Vec<(String, String)>>, anyhow::Error> {
+        Ok(std::collections::HashMap::new())
     }
 
     async fn broadcast_crusher_sampling_policy(
